refactor: Query 관련 패키지 구조 변경 및 메서드 추가
1. Query 관련 클래스의 패키지를 `application.commandquery`에서 `adapter.in.web`으로 이동한다.
2. QueryResponse 클래스에 QueryResult를 기반으로 객체를 생성하는 `from` 메서드 추가한다.
3. QueryRequest 클래스에 QueryCommand로 변환하는 `toCommand` 메서드 추가한다.

Resolves #47
