+++ src/main/java/com/personalproject/llmmanager/querylog/adapter/in/web/request/QueryRequest.java
@@ -1,5 +1,6 @@
-package com.personalproject.llmmanager.querylog.application.commandquery;
+package com.personalproject.llmmanager.querylog.adapter.in.web.request;
 
+import com.personalproject.llmmanager.querylog.application.dtos.command.QueryCommand;
 import jakarta.validation.constraints.NotBlank;
 import org.hibernate.validator.constraints.URL;
 
@@ -11,4 +12,7 @@ public record QueryRequest (
         @NotBlank(message = "모델 타입은 null이거나 비어있을 수 없습니다.")
         String model
 ) {
+    public QueryCommand toCommand() {
+        return new QueryCommand(this.url, this.model);
+    }
 }
