+++ src/main/java/com/personalproject/llmmanager/querylog/adapter/in/web/response/QueryResponse.java
@@ -1,5 +1,6 @@
-package com.personalproject.llmmanager.querylog.application.commandquery;
+package com.personalproject.llmmanager.querylog.adapter.in.web.response;
 
+import com.personalproject.llmmanager.querylog.application.dtos.result.QueryResult;
 import com.personalproject.llmmanager.querylog.domain.model.QueryLog;
 import com.personalproject.llmmanager.users.domain.model.Users;
 
@@ -23,6 +24,16 @@ public record QueryResponse (
                 queryLog.getType().getName(),
                 queryLog.getUsedTokens(),
                 users.getTokens().getRemainingTokens()
+        );
+    }
+
+    public static QueryResponse from(QueryResult queryResult) {
+        return new QueryResponse(
+                queryResult.userName(),
+                queryResult.answer(),
+                queryResult.model(),
+                queryResult.usedToken(),
+                queryResult.remainingToken()
         );
     }
 }
